+++ utils/Util_Getter.py
@@ -11,23 +11,17 @@ from database_operations import USERNAME, PASSWORD, HOSTNAME, PORT, SERVER_DATA
 from database_models import get_tracker
 from database_operations import get_asyncio_db_engine
 from utils.utils import get_guild
+from Generic.Generic_Utilities import Utilities
+from D4e.D4e_utilities import D4e_Utilities
 
 
-async def get_character(char_name, ctx, guild=None, engine=None):
+async def get_utilities(ctx, guild=None, engine=None):
     logging.info("get_character")
     if engine is None:
         engine = get_asyncio_db_engine(user=USERNAME, password=PASSWORD, host=HOSTNAME, port=PORT, db=SERVER_DATA)
     guild = await get_guild(ctx, guild)
-
-    if guild.system == "EPF":
-        return await get_EPF_Character(char_name, ctx, guild=guild, engine=engine)
+    if guild.system == "D4e":
+        return D4e_Utilities(ctx, guild, engine)
     else:
-        tracker = await get_tracker(ctx, engine, id=guild.id)
-        async_session = sessionmaker(engine, expire_on_commit=False, class_=AsyncSession)
-        try:
-            async with async_session() as session:
-                result = await session.execute(select(tracker).where(tracker.name == char_name))
-                character = result.scalars().one()
-            return Character(char_name,ctx, engine, character, guild=guild)
-        except NoResultFound:
-            return None
+        return Utilities(ctx, guild, engine)
+
